♻️ Refactor : 코드 전체 리펙토링
