+++ app/src/main/java/gdsc/solutionchallenge/saybetter/saybetterlearner/ui/component/dialog/InfoProfileDialog.kt
@@ -70,8 +70,5 @@ class InfoProfileDialog {
                         })
             }
         }
-
-
-
     }
 }
