+++ app/src/main/java/gdsc/solutionchallenge/saybetter/saybetterlearner/ui/login/LoginActivity.kt
@@ -75,10 +75,10 @@ class LoginActivity: ComponentActivity(), AuthView {
                 authService.postCommonLogin(AuthCommonRequest("", email.toString(), ""))
                 this.googleIdToken = googleIdToken
             }
-        }
+        }//임시로 제거
 
         setContent {
-            LoginScreen(login = {
+            LoginView(login = {
                 mainRepository.login(testid) { isDone, reason ->
                     if (!isDone) {
                         Log.d("login", "로그인 실패, $reason")
@@ -89,16 +89,6 @@ class LoginActivity: ComponentActivity(), AuthView {
                 }
             })
         }
-    }
-
-    @Preview(widthDp = 1280, heightDp = 800)
-    @Composable
-    fun LoginPreview() {
-        LoginScreen(login = {
-            finish()
-            val intent = Intent(this, MenuActivity::class.java)
-            startActivity(intent)
-        })
     }
 
     override fun onPostLoginSuccess(response: GeneralResponse<AuthCommonResponse>) {
@@ -127,66 +117,5 @@ class LoginActivity: ComponentActivity(), AuthView {
 
     override fun onPostLoginFailure(isSuccess: Boolean, code: String, message: String) {
         TODO("Not yet implemented")
-    }
-}
-@Composable
-fun LoginScreen(login: () -> Unit) {
-    Surface {
-        Box {
-            Column (modifier = Modifier){
-                Row (modifier = Modifier.fillMaxSize()){
-                    Column (modifier = Modifier
-                        .fillMaxWidth(0.45f)
-                        .fillMaxHeight()
-                        .padding(start = 50.dp, top = 80.dp, end = 50.dp)){
-
-                        Text(text = "Say Better Life, Say Better dream",
-                            color = MainGreen,
-                            fontSize = 25.sp)
-
-                        Spacer(modifier = Modifier.height(20.dp))
-
-                        Text(text = "반가워요!\nSay Better를 시작해볼까요?",
-                            fontSize = 40.sp,
-                            fontWeight = FontWeight.W800)
-                    }
-                    Image(painter = painterResource(id = R.drawable.img_login),
-                        contentDescription = null,
-                        modifier = Modifier.fillMaxSize())
-                }
-            }
-            Row (modifier = Modifier
-                .fillMaxSize()
-                .padding(horizontal = 50.dp, vertical = 50.dp),
-                verticalAlignment = Alignment.Bottom){
-                Box (modifier = Modifier
-                    .fillMaxWidth()
-                    .height(70.dp)
-                    .border((1.5).dp, Color.Black, RoundedCornerShape(100.dp))
-                    .background(White)
-                    .clickable(
-                        interactionSource = remember { MutableInteractionSource() },
-                        indication = CustomClickEvent
-                    ) { login() }
-                ){
-                    Row (modifier = Modifier
-                        .fillMaxSize(),
-                        verticalAlignment = Alignment.CenterVertically,
-                        horizontalArrangement = Arrangement.Center){
-
-                        Spacer(modifier = Modifier.width(40.dp))
-
-                        Image(painter = painterResource(id = R.drawable.google_logo),
-                            contentDescription = null,
-                            modifier = Modifier
-                                .size(40.dp))
-
-                        Text(modifier = Modifier.padding(start = 20.dp), text = "Google로 로그인하기",
-                            fontSize = 30.sp,
-                            fontWeight = FontWeight.W600)
-                    }
-                }
-            }
-        }
     }
 }
