+++ app/src/main/java/gdsc/solutionchallenge/saybetter/saybetterlearner/ui/menu/MenuActivity.kt
@@ -51,6 +51,7 @@ import gdsc.solutionchallenge.saybetter.saybetterlearner.ui.info.InfoActivity
 import gdsc.solutionchallenge.saybetter.saybetterlearner.ui.setting.SettingActivity
 import gdsc.solutionchallenge.saybetter.saybetterlearner.ui.theme.White
 import gdsc.solutionchallenge.saybetter.saybetterlearner.ui.videocall.VideoCallActivity
+import gdsc.solutionchallenge.saybetter.saybetterlearner.utils.CustomAlertDialogState
 import gdsc.solutionchallenge.saybetter.saybetterlearner.utils.customclick.CustomClickEvent
 import gdsc.solutionchallenge.saybetter.saybetterlearner.utils.permission.checkAndRequestPermissions
 import gdsc.solutionchallenge.saybetter.saybetterlearner.utils.webrtc.service.MainService
@@ -66,128 +67,56 @@ class MenuActivity: ComponentActivity() , MainService.CallEventListener {
         val onClickCancel: () -> Unit = {},
     )
 
-    private var userid : String? = null
+    private var userId: String? = null
     private val testUser: String = "testUser1"
     private var currentReceivedModel: DataModel? = null
-    val TAG : String = "ServiceDebug"
-
-    var count: Int = 0
+    private val customAlertDialogState = mutableStateOf(TestDialogState())
 
     //Hilt 종속성 주입
     @Inject lateinit var mainRepository : MainRepository
     @Inject lateinit var mainServiceRepository : MainServiceRepository
 
-    private val customAlertDialogState = mutableStateOf(TestDialogState())
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        init()
 
         setContent {
-            MenuPreview()
-        }
-
-        Log.d(TAG, "oncreate")
-        init()
-    }
-
-    @Preview(widthDp = 1280, heightDp = 800)
-    @Composable
-    fun MenuPreview() {
-        var isCaller: Boolean? = null
-
-        val context = LocalContext.current
-
-        /** 요청할 권한 **/
-        val permissions = arrayOf(
-            Manifest.permission.RECORD_AUDIO,
-            Manifest.permission.CAMERA
-        )
-
-        val launcherMultiplePermissions = rememberLauncherForActivityResult(
-            ActivityResultContracts.RequestMultiplePermissions()
-        ) { permissionsMap ->
-            val areGranted = permissionsMap.values.reduce { acc, next -> acc && next }
-            /** 권한 요청시 동의 했을 경우 **/
-            if (areGranted) {
-                Log.d("test5", "권한이 동의되었습니다.")
-                resetDialogState(customAlertDialogState)
-                StartVideoCall(testUser, isCaller!!)
-            }
-            /** 권한 요청시 거부 했을 경우 **/
-            else {
-                Log.d("test5", "권한이 거부되었습니다.")
-            }
-        }
-
-        val menuList = listOf(
-            menu("레벨 테스트", R.drawable.menu_level),
-            menu("그림 상징 의사소통", R.drawable.menu_symbol),
-            menu("텍스트 의사소통", R.drawable.menu_text),
-            menu("AI 챗봇", R.drawable.menu_chatbot),
-            menu("설정", R.drawable.menu_setting),)
-
-        Surface(color = White,
-            modifier = Modifier.fillMaxSize()
-        ){
-            if (customAlertDialogState.value.isClick) {
-                TestDialog().CallTestDialog(
-                    onClickSure = {
-                        isCaller = false
-                        checkAndRequestPermissions(
-                            context,
-                            permissions,
-                            launcherMultiplePermissions,
-                            onPermissionsGranted = {    //권한이 이미 다 있을 때
-                                resetDialogState(customAlertDialogState)
-                                StartVideoCall(currentReceivedModel?.sender!!, isCaller!!)
-                            }
-                        )
-                    },
-                    onClickCancel = { resetDialogState(customAlertDialogState) }
-                )
-            }
-            MenuBar(menuList = menuList,
-                ClickSymbol = {
-                    isCaller = true
-                    checkAndRequestPermissions(
-                        context,
-                        permissions,
-                        launcherMultiplePermissions,
-                        onPermissionsGranted = {    //권한이 이미 다 있을 때
-                            StartVideoCall(testUser, isCaller!!)
-                        }
-                    )
-                    // 권환을 받아야할 때
-                },
-                ClickChatbot = {
+            MenuView(
+                resetDialogState = {resetDialogState(customAlertDialogState)},
+                startVideoCall = {testUser, isCaller -> startVideoCall(testUser, isCaller) },
+                user = userId!!,
+                currentReceivedModel = currentReceivedModel,
+                customAlertDialogState = customAlertDialogState,
+                onClickChatbot = {
                     intent = Intent(this@MenuActivity, ChatBotActivity::class.java)
                     startActivity(intent)
                 },
-                ClickSetting = {
+                onClickSetting = {
                     intent = Intent(this@MenuActivity, SettingActivity::class.java)
                     startActivity(intent)
                 },
-                ClickLevel = {
+                onClickLevel = {
                     intent = Intent(this@MenuActivity, InfoActivity::class.java)
                     startActivity(intent)
-                })
+                }
+            )
         }
+
     }
 
     private fun init(){
-        userid = intent.getStringExtra("userid")
-        if(userid == null) finish()
-
+        userId = intent.getStringExtra("userid")!!
         MainService.listener = this
         startMyService()
     }
 
     private fun startMyService() {
-        mainServiceRepository.startService(userid!!)
+        mainServiceRepository.startService(userId!!)
     }
 
     //Video call 클릭되었을 때
-    private fun StartVideoCall(targetUserid : String, isCaller: Boolean) {
+    private fun startVideoCall(targetUserid : String, isCaller: Boolean) {
         mainRepository.sendConnectionRequest(targetUserid) {
             if(it) {
                 //videocall 시작해야함
@@ -222,81 +151,6 @@ class MenuActivity: ComponentActivity() , MainService.CallEventListener {
         state.value = TestDialogState()
     }
 
-}
-@Composable
-fun MenuBar(menuList : List<menu>,
-            ClickLevel:() ->Unit,
-            ClickSymbol: () ->Unit,
-            ClickChatbot:() ->Unit,
-            ClickSetting:() -> Unit) {
-    Box {
-        Column (modifier = Modifier
-            .fillMaxWidth()
-            .padding(top = 100.dp),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ){
-            Image(painter = painterResource(id = R.drawable.ic_say_better),
-                contentDescription = null,
-                modifier = Modifier
-                    .height(70.dp)
-                    .width(300.dp),)
-        }
-
-        Column(modifier = Modifier.fillMaxSize(),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally) {
-            LazyRow {
-                items(menuList) { menuEntity ->
-                    MenuItem(menuEntity, clickMenu = {
-                        when (menuEntity.title) {
-                            "레벨 테스트" -> {
-                                ClickLevel()
-                            }
-                            "그림 상징 의사소통" -> {
-                                ClickSymbol()
-                            }
-                            "AI 챗봇" -> {
-                                ClickChatbot()
-                            }
-                            "설정" -> {
-                                ClickSetting()
-                            }
-                            else -> {
-                            }
-                        }
-                    })
-                    Log.d("permission", "call FeatureThatRequiresCameraPermission")
-                    if (menuEntity != menuList.last()) Spacer(modifier = Modifier.width(30.dp))
-
-                }
-            }
-        }
-    }
-
-}
-
-@Composable
-fun MenuItem(menuEntity : menu, clickMenu: () -> Unit) {
-    Column (horizontalAlignment = Alignment.CenterHorizontally,
-        modifier = Modifier.clickable (
-            interactionSource = remember{ MutableInteractionSource() },
-            indication = CustomClickEvent
-        ) {
-            clickMenu()
-        }){
-        Image(
-            painter = painterResource(id = menuEntity.img),
-            contentDescription = null,
-            modifier = Modifier
-                .width(200.dp)
-                .height(200.dp)
-                .background(Color.White, shape = RoundedCornerShape(50.dp))
-        )
-        Spacer(modifier = Modifier.height(20.dp))
-        Text(text = menuEntity.title,
-            fontSize = 20.sp,
-            fontWeight = FontWeight.W600)
-    }
 }
