+++ app/src/main/java/gdsc/solutionchallenge/saybetter/saybetterlearner/ui/menu/MenuBar.kt
@@ -1,0 +1,60 @@
+package gdsc.solutionchallenge.saybetter.saybetterlearner.ui.menu
+
+import android.util.Log
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import gdsc.solutionchallenge.saybetter.saybetterlearner.R
+import gdsc.solutionchallenge.saybetter.saybetterlearner.model.data.local.entity.menu
+
+@Composable
+fun MenuBar(
+    menuList: List<MenuItemModel>,
+    onClickMenuItem: (MenuItemModel) -> Unit
+) {
+    Box {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(top = 100.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Image(
+                painter = painterResource(id = R.drawable.ic_say_better),
+                contentDescription = null,
+                modifier = Modifier
+                    .height(70.dp)
+                    .width(300.dp)
+            )
+        }
+
+        Column(
+            modifier = Modifier.fillMaxSize(),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            LazyRow {
+                items(menuList) { menuItem ->
+                    MenuItem(menuItem) {
+                        onClickMenuItem(menuItem)
+                    }
+                    if (menuItem != menuList.last()) Spacer(modifier = Modifier.width(30.dp))
+                }
+            }
+        }
+    }
+}
