+++ app/src/main/java/gdsc/solutionchallenge/saybetter/saybetterlearner/ui/menu/MenuItemModel.kt
@@ -1,0 +1,7 @@
+package gdsc.solutionchallenge.saybetter.saybetterlearner.ui.menu
+
+data class MenuItemModel(
+    val title: String,
+    val iconRes: Int,
+    val onClick: () -> Unit
+)
